requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain-community
requests
python-dotenv
fastapi
uvicorn
python-dotenv
google-generativeai
